requirements.txt
fastapi>=0.103.0,<1.0.0
uvicorn[standard]>=0.23.0,<1.0.0
httpx>=0.24.1,<1.0.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.1,<3.0.0
python-dotenv>=1.0.0,<2.0.0
redis>=5.0.0,<6.0.0
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<0.24.0
streamlit>=1.28.0,<2.0.0
pillow>=10.0.0,<11.0.0
